TY-2349 Reading assets manifest before engine initialization (#108)
[Jira ref](https://xainag.atlassian.net/browse/TY-2349)

This PR fixes the flutter plugin & example app, by moving reading of the assets `Manifest` file outside of the engine.

`PlatformAssetBundle` which is used used to read the manifest file from the flutter assets, can't be used in an isolate other than main isolate. We also need to use `PlatformAssetBundle` to read the file, cause Flutter doesn't support the `Isolate.Isolate.resolvePackageUri` method. So easiest solution is to read the file first and pass it with `Configuration` object during engine initialisation.