TY-2391 Take model assets from app configuration (#133)
[Jira ref](https://xainag.atlassian.net/browse/TY-2391)

In order to use the AI assets bundled with the application, we need to copy them first to a directory accessible by the worker isolate, because `rootBundle` can't be used in an isolate (other then the main one). This PR adds a `FlutterAiAssetCopier` to copy the AI assets from the app bundle to a directory used by the `DataProvider`.

The other option would be to move the code used to retrieve the AI assets entirely to the main isolate and just pass the `SetupData` with the `Configuration` during engine init. This would require a refactor (maybe not a big issue, depends also a bit on needed tests adjustments), but also then limiting ourselves to use the assets only from the bundle replacing the `HttpAssetFetcher`.

This way (implemented by this PR) we can start with assets bundled with the app, but then fetch updated assets on the fly over the network if needed, without the need to release a new version of the app.