TY-2409 ChangedDocumentsReporter (#153)
[Jira ref](https://xainag.atlassian.net/browse/TY-2409)

This PR adds a `ChangedDocumentsReporter` class to manage a stream of `DocumentsUpdated` engine event, which is sent when `Document` entities that are exposed to the app change their internal state. This event acts as a notification that something has updated, so that the app can replace changed documents in the state that it manages.

For now this notyfication will happen only in response to `UserReactionChanged` event, and will be extended if needed for other events.