Fix domain names being treated as valid URLs
In the Newscatcher API, the field `clean_url` actually contains only a domain name. This led to some confusion regarding how to treat this value, and we were trying to parse it as a URL in `Document`. We're thus renaming it to make clearer what this field actually contains and prevent future confusion.