TY 2483 Fix domain names being treated as valid URLs (#176)
In the Newscatcher API, the field `clean_url` actually contains only a domain name. This led to some confusion regarding how to treat this value, and we were trying to parse it as a URL in `Document`. We're thus renaming it to make clearer what this field actually contains and prevent future confusion.

Co-authored-by: Lucas Jenss <lucas.jenss.external@xayn.com>
Co-authored-by: Andrea Corradi <andrea.corradi@xayn.com>