TY-2497 - Fix errors when deserializing articles with null fields
[TY-2497]

We discovered that, contrary to their documentation, Newscatcher responses
sometimes contain articles which are missing non-optional fields, causing
our deserializing to fail. As a remedy, we're introducing defaults for all
fields which may potentially exhibit this behaviour. In a follow-up, we may
need to introduce additional filtering to discard deserialized articles which
are missing crucial information (such as the excerpt).