TY-2523 do more requests (#273)
Ticket:
- [TY-2523]

Summary:
- make more requests if we don't have `min_articles` of articles. the number of requests is limited by `max_requests`
- added tests

During the implementation, I thought about whether the whole thing could also be implemented with `tower`. Writing the layers will require some thinking but long term I think we would benefit from it as it is more flexible. I saw that `kube` rebuild their client code with tower so there is a good reference that we can use. It would be nice if we could do a small poc that implements this pr in form of a tower service.
 


[TY-2523]: https://xainag.atlassian.net/browse/TY-2523?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ