TY-2531 Tool to create mock model (#282)
[TY-2531 - Tool to create mock model](https://xainag.atlassian.net/browse/TY-2531)
[TY-2538 - Change test to use mock models](https://xainag.atlassian.net/browse/TY-2538)

This PR contains a Python script that generates dummy `bert` or `smbert` onnx models which have the same inputs and outputs as the models we are using in production.

I've generated these two models using that script and added them to version control in the assets folder of our flutter example. Those assets are then used by our integration tests. I've adjusted a helper method which prepares the asset `Manifest` so that it contains mocked versions of the models instead of real ones.