TY-2705  Process incoming events sequentially in DiscoveryEngineWorker (#294)
[Task - process events sequentially](https://xainag.atlassian.net/browse/TY-2705)

This PR adds an additional stream in the `DiscoveryEngineWorker` for incoming messages. The purpose of that stream is to force sequantial processing of events, so that the previous event is fully processed before the next one starts. This is achieved simply by adding [`asyncMap`](https://api.dart.dev/stable/2.16.2/dart-async/Stream/asyncMap.html) method to the stream.