TY 2690 move xayn ai (#297)
We move the relevant part of the xayn_ai project in this repository under `discovery_engine_core/ai/`.
The code under `xayn-ai/reranker` has been removed and with it all the parts that were not used anymore.
Most of removed code was under `xayn-ai`, there could still be code in the other crate that is unused but marked as `pub` so the compiler don't tell us. Looking for this code is out of the scope of this PR.

I left some comment on the parts that are important to review, a special look should be give to the second commit "Adapt path to assets" that modify the path to reach the assets in this repository, this is the only big change in the imported code.

The first commit import the crates, the second fix the path to reach the assets to run the tests for the parts that we are interested in. After this we have commits that removes components that we don't want to use anymore followed by commits that remove code that the compiler tell us that is unused.