Handle stacks that are not ready to ask new data (#335)
If a stack is not ready to return data we don't count it as successful or failed stacks.

At the moment if `TrustedSources` has no  sources and is the only stack that required to be updated we are returning an error but we should not.

This PR allows a stack to return `NotReady` as a variant of the error, if a stack is not ready we exclude it from the count of stacks w.r.t. the return of errors.

Having `new_items` returning something like
```rust
enum NewItems {
  NotReady,
  Items(Vec<Article>),
  Err(GenericError),
}
```
would be nicer but for how we are handling errors in the stack and engine it will require more changes.