TY-2802 improve error handling for invalid engine state (#361)
Ticket: 
- [TY-2802]

Summary:

Currently, if the engine cannot be restored from the saved state (if the version of the saved state > the expected version), the error is simply propagated to the caller. However, the caller has no way of clearing the invalid engine state or starting the engine from scratch. Therefore, we need to improve the error handling for this specific error. The idea is to just clear invalid engine status when this error occurs and return an error to the caller saying that they should try again. 

The improved error handling is implemented in Dart so that team blue can catch those errors and emit events for them. 

[TY-2802]: https://xainag.atlassian.net/browse/TY-2802?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ