TY 2873 Fetch from all market before update the stack (#420)
This PR fetch articles from all the markets before updating the stack.
The problem with the current code is that if updating from a market can bring the stack to a state where it does not need to be updated. This means that when we try to update the other markets this stack will be skipped.