More strict justfile (#503)
Now the `rust-check` task can pass even if `cargo clippy` returns an error. This is because we are using `;` as a command separator and we didn't set `-e` as a bash option.

This PR set `-euxo pipefail` for the main shell and all the shell that we use to make it fail early.