TO-3248 Subtract one time for each stack when not ready (#547)
[TO-3248]

At the moment we are subtracting `1` from `ready_stacks` each time a stack is not ready multiplied for the number of market.
This is not correct and we should subtract `1` only once. At the moment if a stack is not ready for a market it will not be ready for other markets and no other property that can change the readiness can be updated during `update_stacks`.

[TO-3248]: https://xainag.atlassian.net/browse/TO-3248?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ