ET-3376 use futures unordered (#637)
Use `FuturesUnordered` instead of `join_all`.

I'm still not sure it makes any sense as it could easily be a neglible small performance reduction instead of a similar small increase and it either way is IMHO a pre-mature optimization. Just to put it into context `join_all` will internally use `FuturesUnordered` (through `FuturesOrdered`) iff and only iff there are "many" futures (> 30), else it will not use it as it's slower.

There is also the thing that tokio promotes using `spawn` when you join over multiple futures. E.g. tokio has some mechanism to auto-brake live locks but that same mechanism can cripple larger nested futures, `FuturesUnordered` somewhat works around it but there are limits in how much it can do so. I might be better to spawn the requests and then just use `join_all` for simplicity (simpler as in simpler for us).  One thing which speaks against doing this is that spawing futures requires them to be static, so we would need to sprinkle in some `Arc`s.

**References:**

- [ET-3376]

[ET-3376]: https://xainag.atlassian.net/browse/ET-3376?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ