Split ai crate (#710)
Split out the coi system. Simplifications or other changes are out of scope.
    
The aim is to separate the code that the web api needs from the one that only the engine needs.
The code related to the engine will be deleted in the future.


Since the code left in `ai/ai`  will be deleted it does not need a thorough review.