Initial validation and linting of the openapi specification (#722)
Adds an initial linter for the openapi specification.
The linter has been added to the docker image.
A just file command has also been added. Currently, that command is not called during `just pp` since it requires to have `node.js` installed. We can talk about if we want this also there or not.

Lots of lints are currently disabled. Before fixing them, we need to refactor the specification to remove the duplication that we have. Once that is done, we can enable more lints. In the meanwhile, we at least have some sanity checks for the specification.