Don't consider partial files as downloaded (#775)
The current download script download an asset directly in the name we give. If the connections drops that we think that the file has been downloaded. This PR fix this by downloading to a temporary file, this also allows curl to resume the download from the previous state instead of starting from scratch.