ET-4004 Return bad request if document does not exists (#827)
Return an error if some of the documents do not exist.
The documentation was mentioning that the endpoint does not return an error and then contradicts itself by stating that it could return 400. The part around the idempotency of the method was also wrong. Being idempotent does not mean that a method does not return an error if something does not exist.
The HTTP standard requires DELETE methods to be idempotent and it specify that it is in respect to the state of the server not the response. 
```
A request method is considered "idempotent" if the intended EFFECT ON THE SERVER of multiple identical requests with that method is the same as the effect for a single such request. Of the request methods defined by this specification, PUT, DELETE, and safe request methods are idempotent.
```

Returning an error makes `DELETE /documents/<id>` behaving like `GET /documents/<id>`  where we return 400 if the document does not exists.