Unify release build (#893)
Unify the two ci to build the image of the service in one. Remove unused workflow to build other images.
When you create an image both services will be built, they will have a different image name but the same tag.
The workflow allows you to specify:
 * platform to build for,
 * model name and version to use in the image

It automatically creates a tag that describes what is in the image, when it was created and from which commit. It still allows overriding the default with a custom tag.

Co-authored-by: Andrea Corradi <andrea.corradi@xayn.com>
Co-authored-by: Simon Joecks <simon.joecks@gmail.com>