ET-4531 Use xaynia2 as default model (#957)
Use Xaynia 2 as the default model of the model.
Unfortunately changing the token size will be a breaking change for some deployments so it cannot be done in this PR but we need to wait for each deployment to configure that explicitly. Later we should also remove defaults that can easily break backward compatibility.

Co-authored-by: Andrea Corradi <andrea.corradi@xayn.com>