Remove obsolete migration test
All the deploy we have pass that migration point so this test is
useless.

Pointing to our version in the past increase the compilation time for
the tests by 30s.
This also create problem with cargo chef that is not able to handle
a self dependency on an older version.